Move retention offer evaluation out of the server action module
Next.js only allows async exports from "use server" files; the synchronous
buildRetentionOffers helper broke the Vercel build.
